requirements.txt
colored
deeplake
langchain
openai
python-dotenv
tiktoken
